solved '*' problem
the feedback did not return '*' when the digit in the guess had the same
position as that in the code in cases where the digit was repeating in
the guess. this was found to be due to the use of the string method
index which returns the index of the first occurance of the element.
this was fixed by declaring a variable named index and updating the
value in each iteration of the while loop